🐛 fix: mirror the notification deep merge in the optimistic store cache
