+++ packages/database/src/models/workspaceUserSettings.ts
@@ -1,4 +1,5 @@
-import type { NotificationSettings, WorkspaceUserPreference } from '@lobechat/types';
+import type { WorkspaceUserPreference } from '@lobechat/types';
+import { mergeNotificationSettings } from '@lobechat/utils/mergeNotificationSettings';
 import { and, eq } from 'drizzle-orm';
 
 import { workspaceUserSettings } from '../schemas/workspace';
@@ -18,42 +19,6 @@ import type { LobeChatDatabase } from '../type';
  * pair upserts, so members who never customize anything simply have no row
  * and callers fall through to defaults on read.
  */
-/**
- * Merge a notification-settings patch over the stored bag without dropping
- * sibling keys: clients patch a single switch at a time (one channel's
- * `enabled`, or one `items[category][type]` leaf), so a shallow replace at
- * any level would wipe this member's other toggles for that channel.
- */
-const mergeNotificationSettings = (
-  current: NotificationSettings | undefined,
-  patch: NotificationSettings,
-): NotificationSettings => {
-  const next: NotificationSettings = { ...current };
-  for (const [channel, channelPatch] of Object.entries(patch)) {
-    if (!channelPatch) continue;
-    const key = channel as keyof NotificationSettings;
-    const currentChannel = next[key];
-    const mergedItems: Record<string, Record<string, boolean>> = {
-      ...currentChannel?.items,
-      ...channelPatch.items,
-    };
-    const items = channelPatch.items
-      ? Object.fromEntries(
-          Object.keys(mergedItems).map((category) => [
-            category,
-            { ...currentChannel?.items?.[category], ...mergedItems[category] },
-          ]),
-        )
-      : currentChannel?.items;
-    next[key] = {
-      ...currentChannel,
-      ...channelPatch,
-      ...(items ? { items } : {}),
-    };
-  }
-  return next;
-};
-
 export class WorkspaceUserSettingsModel {
   private readonly db: LobeChatDatabase;
   private readonly userId: string;
