+++ src/store/user/slices/workspaceUserSettings/action.test.ts
@@ -83,4 +83,36 @@ describe('WorkspaceUserSettingsActionImpl', () => {
       selected: false,
     });
   });
+
+  it('optimistically deep-merges one notification switch without dropping sibling toggles', async () => {
+    const state = {
+      workspaceUserPreference: {
+        notification: {
+          email: { items: { workspace: { workspace_member_joined: false } } },
+          inbox: { enabled: false },
+        },
+      },
+    };
+    const set = vi.fn((patch: Partial<typeof state>) => Object.assign(state, patch));
+    const action = new WorkspaceUserSettingsActionImpl(set as never, () => state as never);
+    vi.spyOn(workspaceUserSettingsService, 'updatePreference').mockResolvedValue();
+
+    // Patch a single other leaf — the earlier email item toggle and the inbox
+    // master switch must both survive in the optimistic cache, mirroring what
+    // the server-side deep merge keeps in the DB.
+    await action.updateWorkspaceUserPreference({
+      notification: {
+        email: { items: { workspace: { workspace_payment_failed: false } } },
+      },
+    });
+
+    expect(state.workspaceUserPreference.notification).toEqual({
+      email: {
+        items: {
+          workspace: { workspace_member_joined: false, workspace_payment_failed: false },
+        },
+      },
+      inbox: { enabled: false },
+    });
+  });
 });
